move mock helper and exclude it from tsconfig
