+++ packages/insomnia-testing/src/test-helpers/send-request-mock.ts
@@ -1,6 +1,6 @@
 import { mocked } from 'ts-jest/utils';
 
-import { SendRequestCallback } from './run/insomnia';
+import { SendRequestCallback } from '../run/insomnia';
 
 export const mockedSendRequest = <TResp extends {status: number}>(response?: TResp) => {
   return mocked<SendRequestCallback<TResp>>(jest.fn()).mockResolvedValue(response || { status: 200 } as TResp);
